Fix CI failures from first full-workflow push.
Cast seccomp allow/deny to uint32 for linux/amd64, bump golang.org/x/crypto
for Trivy HIGH findings, correct the gitleaks module path, and silence
shellcheck SC2035 on digest scripts.

+++ internal/confine/apply_linux.go
@@ -244,8 +244,8 @@ func seccompDenyEngineering(denyNet bool) error {
 		offNR   = 0
 		offArch = 4
 	)
-	deny := unix.SECCOMP_RET_ERRNO | uint32(unix.EPERM)
-	allow := unix.SECCOMP_RET_ALLOW
+	deny := uint32(unix.SECCOMP_RET_ERRNO) | uint32(unix.EPERM)
+	allow := uint32(unix.SECCOMP_RET_ALLOW)
 	ldAbs := uint16(unix.BPF_LD | unix.BPF_W | unix.BPF_ABS)
 	jmpJEQ := uint16(unix.BPF_JMP | unix.BPF_JEQ | unix.BPF_K)
 	retK := uint16(unix.BPF_RET | unix.BPF_K)
